streamlit/requirements.txt
streamlit==1.28.1
pandas==2.1.1
plotly==5.17.0
folium==0.14.0
streamlit-folium==0.15.0
requests==2.31.0
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
minio>=7.1.17
sqlalchemy==2.0.23